+++ index.js
@@ -3,6 +3,7 @@
 module.exports = {
   rules: {
     'no-class': require('./rules/no-class'),
+    'no-delete': require('./rules/no-delete'),
     'no-let': require('./rules/no-let'),
     'no-loops': require('./rules/no-loops'),
     'no-mutating-assign': require('./rules/no-mutating-assign'),
@@ -15,6 +16,7 @@ module.exports = {
       rules: {
         'no-var': 'error',
         'fp/no-class': 'error',
+        'fp/no-delete': 'error',
         'fp/no-let': 'error',
         'fp/no-loops': 'error',
         'fp/no-mutating-assign': 'error',